+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Support for object responses
 - The `Search` component no longer requires an explicit wrapper node for its children
 - Configurable prefix
-- Added InfiniteScroll component
+- Added `InfiniteScroll` component
+- Added debounce support to the `Param` component
 
 ## 1.1.0
 - The fields used for the label and value of Checkbox/Radio/Select components